refactor(root): move Index and InstantSearch to widgets
