+++ packages/react-instantsearch-core/src/widgets/Index.tsx
