+++ packages/react-instantsearch-core/src/widgets/InstantSearch.tsx
