+++ packages/react-instantsearch-core/src/widgets/__tests__/Index.js
